# refactor(perf): improve tag serialization